refactor: imports and style
Story: [ch77]
