+++ packages/server/http-server/src/domain/usecases/errors/index.ts
@@ -1,2 +1,3 @@
-export * from '@usecases/errors/account-already-exists';
-export * from '@usecases/errors/user-validation';
+export * from './account-already-exists';
+export * from './user-validation';
+export * from './document-already-in-use';
