+++ packages/server/http-server/src/domain/usecases/output-ports/repositories/users/index.ts
@@ -1,1 +1,2 @@
-export * from '@usecases/output-ports/repositories/users/users';
+export * from './users';
+export * from './fake';
